API Reference

Index

Iterators emitting only the grouped elements or indices

GroupNumbers.groupby2 — Method

groupby2(xs; keyfunc=identity, compare=isequal)

An iterator that yields a group of elements xg from the iterator xs, where the key is computed by applying keyfunc to each element of xs. Compare adjacent keys by compare function, then, emits the vector xg of the grouped elements.

See documentation

GroupNumbers.groupby2_indices — Method

groupby2_indices(xs; keyfunc=identity, compare=isequal)

An iterator that yields a group of indices ig from the iterator xs, where the key is computed by applying keyfunc to each element of xs. Compare adjacent keys by compare function, then, emits the vector ig of the grouped indices.

See documentation

GroupNumbers.groupby_numbers — Method

groupby_numbers(xs; keyfunc=identity, kwargs)

An iterator that yields a group of numbers xg from the iterator xs of presumably numbers, where the key is computed by applying keyfunc to each element of xs. Compare adjacent keys by isapprox function with supplied kwargs optional parameters, then, emit the vector xg of the grouped numbers.

See documentation

GroupNumbers.groupby_numbers_indices — Method

groupby_numbers_indices(xs; keyfunc=identity, kwargs)

An iterator that yields a group of indices ig from the iterator xs of presumably numbers, where the key is computed by applying keyfunc to each element of xs. Compare adjacent keys by isapprox function with supplied kwargs optional parameters, then, emits the vector ig of the grouped indices.

See documentation

Iterators emitting keys and grouped elements or indices

GroupNumbers.groupby2_dict — Method

groupby2_dict(xs; keyfunc=identity, compare=isequal)

An iterator that yields a key-and-group pair (k,xg) from the iterator xs, where the key k is computed by applying keyfunc to each element of xs. Compare adjacent keys by compare function, then, emits the key k and the vector xg of the grouped elements .

See documentation

GroupNumbers.groupby2_dict_indices — Method

groupby2_dict_indices(xs; keyfunc=identity, compare=isequal)

An iterator that yields a key-and-group pair (k,ig) from the iterator xs, where the key k is computed by applying keyfunc to each element of xs. Compare adjacent keys by compare function, then, emits the key k and the vector ig of the grouped indices.

See documentation

GroupNumbers.groupby_numbers_dict — Method

groupby_numbers_dict(xs; keyfunc=identity, kwargs)

An iterator that yields a key-and-group pair (k,xg) from the iterator xs of presumably numbers, where the key k is computed by applying keyfunc to each element of xs. Compare adjacent keys by isapprox function with supplied kwargs optional parameters, then, emits the key k and the vector xg of grouped elements.

See documentation

GroupNumbers.groupby_numbers_dict_indices — Method

groupby_numbers_dict(xs; keyfunc=identity, kwargs)

An iterator that yields a key-and-group pair (k,ig) from the iterator xs of presumably numbers, where the key k is computed by applying keyfunc to each element of xs. Compare adjacent keys by isapprox function with supplied kwargs optional parameters, then, emits the key k and the vector ig of the grouped indices.

See documentation